split stories and expose events
